Refactor: Move shelf upload batch operation into UserSyncRepository
- Added `uploadAllSyncedUsersToShelf` to `UserSyncRepository` and `UserRepositoryImpl`.
- Relocated concurrency control (`Semaphore(5)`, `supervisorScope`, `async`, `awaitAll`) from `UploadToShelfService` to `UserRepositoryImpl`.
- Ensured UI interaction (`listener.onSuccess`) and main-thread switching remains in `UploadToShelfService` so the repository is UI-free.
